chore: test merge test cases
